refact: rename config.ts to index.ts and update affected import link
